Move mixed-dtype param group test into dist_checkpointing test_optimizer
The regression test is a dist_checkpointing save/load round-trip, and
tests/unit_tests/dist_checkpointing/test_optimizer.py is where the
sibling keep_in_fp32 coverage already lives (Fp32MarkedModel and
test_float16_optimizer_with_fp32_marked_params, which exercise the
non-distributed Float16OptimizerWithFloat16Params pairing). Co-locating
MixedDtypeNet next to Fp32MarkedModel keeps the feature family's
checkpoint coverage in one file, drops the cross-package import of the
dist_checkpointing test infra, and puts the test in that file's
dedicated CI bucket.

Conform to the file's conventions while moving: class with
teardown_method calling Utils.destroy_model_parallel() instead of
try/finally, and rely on Utils.initialize_model_parallel() for process
group init instead of an explicit _init_distributed() call. Test logic
is unchanged.
